core/requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.0.4
billiard==4.2.0
cattrs==23.2.3
celery==5.3.4
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==41.0.7
cssselect2==0.7.0
daphne==4.0.0
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.5
django-js-asset==2.2.0
django-mptt==0.15.0
django-rest-framework==0.1.0
django-schema-graph==3.1.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
exceptiongroup==1.2.0
freetype-py==2.3.0
html5lib==1.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.5
lxml==4.9.4
oscrypto==1.3.0
Pillow==10.0.1
prompt-toolkit==3.0.43
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.0
pycodestyle==2.11.1
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pyOpenSSL==23.3.0
pypdf==3.17.4
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.32.0
reportlab==4.0.8
requests==2.31.0
rlPyCairo==0.3.0
rpds-py==0.15.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ua-parser==0.18.0
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.1.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.13
zope.interface==6.1
